app/backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
twisted==24.7.0
txaio==23.1.1
typing-extensions==4.12.2
urllib3==2.2.2
zope.interface==7.0.3
psycopg
django-redis==5.0.0
channels_redis
